Fixed order POST
Fixed problems with the API, refactored the code  to separe this action from other components
