+++ app/src/main/java/com/example/motomoapp/models/FoodItem.kt
@@ -13,8 +13,7 @@ data class FoodItem (
     val name: String,
     @SerializedName("dsc") val description: String,
     var price: String,
-    @SerializedName("img") var idImage: String,
-    var url: String = ""
+    @SerializedName("url") var idImage: String,
 )  : Parcelable {
     constructor(parcel: Parcel) : this(
         parcel.readString()!!,
