+++ app/src/main/res/layout/activity_item_detalle.xml
@@ -43,7 +43,7 @@
                 android:adjustViewBounds="true"
                 android:elevation="1dp"
                 android:scaleType="centerCrop"
-                app:imageUrl="@{foodItem.url}"
+                app:imageUrl="@{foodItem.idImage}"
                 android:translationZ="-1dp"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintHorizontal_bias="0.0"
