+++ app/src/main/res/layout/food_item_card.xml
@@ -41,7 +41,7 @@
                 android:layout_height="100dp"
                 android:adjustViewBounds="true"
                 android:scaleType="centerCrop"
-                app:imageUrl="@{foodItem.url}"
+                app:imageUrl="@{foodItem.idImage}"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
                 app:layout_constraintTop_toTopOf="parent" />
